my-company/backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.2
django-ip-geolocation==1.6.1
django-ratelimit==4.1.0
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
promise==2.3
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-memcached==1.62
pywin32==311
redis==7.0.1
requests==2.32.5
rest-framework-simplejwt==0.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
whitenoise==6.11.0
